新增錯誤port警告

+++ app/src/main/kotlin/com/eria/application/main/App.kt
@@ -44,8 +44,7 @@ class MyView : View() {
                         response.body = "test123"
                         response
                     }
-                    http.star(portField.text.toInt())
-
+                         http.star(portField.text)
                 }
             }
             button(" 停止") {
